Move queue code from service object to separate class
